fixes for the spatial object handling
* fixed param spatial_extent handling on get_sample
* fixes on transforming sf objects from and to GeoJSON
* removed suggested connection to package geojsonsf which can also be
handled by sf directly

+++ R/argument_types.R
@@ -1577,22 +1577,16 @@ GeoJson = R6Class(
         warnings("Package sf is not installed but required for GeoJson support.")
       }
       
-      if (is.list(value)) {
+      if (is.list(value) && "type" %in% names(value) && !any(c("sf","sfc") %in% class(value))) {
         tryCatch({
           tmpfile = tempfile()
           jsonlite::write_json(value,tmpfile, auto_unbox=TRUE)
-          value = sf::read_sf(tmpfile, driver="geojson")
+          value = sf::read_sf(tmpfile)
         }, finally = unlink(tmpfile)) 
       }
       
-      if (isNamespaceLoaded("sf")) {
-        if (all(c("XY","POLYGON") %in% class(value))) {
-          value = sf::st_sfc(value)
-        }
-        
-        if (any(c("sf","sfc") %in% class(value))) {
-          value = sf::st_transform(vale,4326)
-        } 
+      if (all(c("XY","POLYGON") %in% class(value))) {
+        value = sf::st_sfc(value)
       }
       
       private$value = value
@@ -1603,15 +1597,21 @@ GeoJson = R6Class(
   ),
   private = list(
     typeCheck = function() {
-      if (is.list(private$value)) {
-        #if list we assume that the geojson object was created as list
-        return(NULL)
-      } 
+       
       
       if ("sf" %in% class(private$value)) {
         return(NULL)
       } else if ("sfc" %in% class(private$value)) {
         return(NULL)
+      } else if (is.list(private$value)) {
+          
+          if (!"type" %in% names(private$value)) {
+            #TODO better probing
+            stop("Value is not GeoJSON.")
+          }
+          
+          #if list we assume that the geojson object was created as list
+          return(NULL)
       } else {
         stop("Class ",paste(class(private$value)), " not supported in GeoJson argument")
       }
@@ -1619,15 +1619,19 @@ GeoJson = R6Class(
       
     },
     typeSerialization = function() {
-      if (is.list(private$value) && all(c("type","coordinates") %in% names(private$value))) {
+      if (any(c("sf","sfc") %in% class(private$value))) {
+        
+        # TODO decide if this is a good idea or whether we are allowed to use crs in GeoJSON for the openEO back-ends
+        value = sf::st_transform(private$value,4326)
+        
+        tryCatch({
+          t = tempfile()
+          write_sf(value,t,driver="geojson")
+          return(jsonlite::read_json(t,simplifyVector = FALSE))
+        }, finally = unlink(t))
+          
+      } else if (is.list(private$value) && "type" %in% names(private$value)) {
         return(private$value)
-      } else if (any(c("sf","sfc") %in% class(private$value))) {
-        if (!.is_package_installed("geojsonsf")) {
-          stop("Package 'geojsonsf' is required for serializing geometries into GeoJson. Please install the package.")
-        } else {
-          gson = geojsonsf::sf_geojson(private$value)
-          return(jsonlite::fromJSON(gson, simplifyVector = FALSE))
-        }
       } else {
         stop("Unsupported value type.")
       }
